requirements.txt
asgiref==3.5.2
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.6
channels==3.0.4
chardet==3.0.4
charset-normalizer==2.0.4
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-dotenv==1.4.2
django-environ==0.10.0
django-extensions==3.2.0
django-filter==22.1
django-registration==3.1.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-test-plus==2.2.0
djangorestframework==3.13.1
freezegun==1.2.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
mysqlclient==2.1.1
oauthlib==3.1.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.3.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.25.9
uWSGI==2.0.21
wrapt==1.14.1
zope.interface==5.4.0
